aws_GHZ/requirements.txt
amazon-braket-sdk
numpy